chatback/pyproject.toml
[build-system]
requires = [
    "setuptools>=78.1.0",
    "wheel>=0.45.1",
]
build-backend = "setuptools.build_meta"

[project]
name = "chatback"
version = "0.6.1"
description = "Backend API for the chat application"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=2.0.40",
    "psycopg2>=2.9.10",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "passlib>=1.7.4",
    "bcrypt>=4.3.0",
    "python-multipart>=0.0.20",
    "httpx>=0.28.1",
    "alembic>=1.15.2",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "scipy>=1.15.2",
    "scikit-learn>=1.6.1",
    "qdrant-client>=1.13.3",
    "tiktoken>=0.9.0",
    "openai>=1.75.0",
    "langchain>=0.3.23",
    "langchain-openai>=0.3.14",
    "langchain-community>=0.3.21",
    "langchain-core>=0.3.54",
    "langchain-text-splitters>=0.3.8",
    "tenacity>=9.1.2",
    "redis>=5.2.1",
    "aioredis>=2.0.1",
    "pyjwt>=2.10.1",
    "cryptography>=44.0.2",
    "email-validator>=2.2.0",
    "argon2-cffi>=23.1.0",
    "termcolor>=3.0.1",
    "PyYAML>=6.0.2",
    "langgraph>=0.3.31",
    "pytz>=2025.2",
    "requests>=2.32.3",
    "Jinja2>=3.1.6",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "coverage>=7.8.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "deptry>=0.23.0",
]

[tool.setuptools]
packages = [
    "app",
]

[tool.black]
line-length = 88
target-version = [
    "py313",
]
include = "\\.pyi?$"

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
overrides = [
    { module = "tests.*", disallow_untyped_defs = false, disallow_incomplete_defs = false },
]
